config, metrics changes

+++ src/cpp/src/metric.hpp
@@ -3,7 +3,7 @@
 #include <cstddef>
 
 struct Metric {
-  size_t id;
+  long id;
   double val;
   std::chrono::time_point<std::chrono::system_clock, std::chrono::nanoseconds>
       timestamp;
